requirements.txt
asyncpg
qdrant-client
python-dotenv
tiktoken
openai
pydub
sounddevice
langchain